refactoring json svc registry to handle yaml files
